Use dynamic import for test-reviewer
