+++ package.json
@@ -12,7 +12,7 @@
     "validate": "npm run create-all-tests -- --verbose --validate",
     "cleanup": "rimraf build",
     "create-all-tests": "cross-var node scripts/create-all-tests --testplan=$npm_config_testplan",
-    "review-tests": "cross-var node --experimental-modules scripts/test-reviewer/index.mjs --testplan=$npm_config_testplan",
+    "review-tests": "cross-var node scripts/test-reviewer --testplan=$npm_config_testplan",
     "update-reference": "node scripts/update-reference.js",
     "make-v2": "node scripts/v2maker.js",
     "test": "jest --runInBand",
